Splitting UIImageView+Extensions and UIImageView+Blavatar
